Variable type has been changed.

+++ saga-choreography/inventory-service/src/main/java/com/market/inventory/application/ports/in/FindInventoryByProductIdInputPort.java
@@ -4,5 +4,5 @@ import com.market.inventory.application.core.domain.Inventory;
 
 public interface FindInventoryByProductIdInputPort {
 
-    Inventory find(Integer productId);
+    Inventory find(Long productId);
 }
